Update libwebp with patch for WebP-Animation.
Bug id for this change is 10512287

Omitted one libwebp patch that was added to Chrome
- f626fe2 Detect canvas and image size mismatch in decoder.
This patch added stricter check for matching the image width & height
and required more data to parse the header.
Since in Android, we use 64 byte buffer to parse the header, this change
was returning NOT_ENOUGH_DATA error while parsing WebP header.

Change-Id: Id73356a85ce1ceeeb822c1c428bf94b7127c6d0f